requirements.txt
pika==1.0.1
PyYAML==5.1
